requirements.txt
fastapi[standard]
sqlalchemy
uvicorn[standard]
celery
pydantic
passlib[bcrypt]
PyJWT
python-dotenv
requests
